refactor(imageframe): update ImageFrame and MainTab to use new structure
