+++ src/components/Body/Tabs/MainTab/ImageFrame/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ImageFrame";
